+++ sbe-tool/src/test/cpp/MessageBlockLengthTest.cpp
@@ -186,7 +186,7 @@ TEST_F(MessageBlockLengthIrTest, shouldHandleAllEventsCorrectlyInOrder)
     std::uint64_t blockLength = headerDecoder.getBlockLength(m_buffer);
 
     const std::size_t result = OtfMessageDecoder::decode(
-        messageBuffer, length, actingVersion, blockLength, messageTokens, *this);
+        messageBuffer, length, actingVersion, static_cast<std::size_t>(blockLength), messageTokens, *this);
     EXPECT_EQ(result, static_cast<std::size_t>(54 - MessageHeader::encodedLength()));
 
     EXPECT_EQ(m_eventNumber, 7);
